requirements.txt
vosk>=0.3.45
pyaudio>=0.2.13
pyserial>=3.5
dash>=2.9.0
dash-bootstrap-components>=1.4.0
plotly>=5.15.0
numpy>=1.24.0
